+++ packages/algoliasearch/src/builds/node.ts
@@ -191,6 +191,16 @@ import {
   waitTask,
 } from '@algolia/client-search';
 import { createNullLogger } from '@algolia/logger-common';
+import {
+  getFrequentlyBoughtTogether,
+  getLookingSimilar,
+  getRecommendations,
+  getRecommendedForYou,
+  getRelatedProducts,
+  getTrendingFacets,
+  getTrendingItems,
+  WithRecommendMethods,
+} from '@algolia/recommend';
 import { Destroyable } from '@algolia/requester-common';
 import { createNodeHttpRequester } from '@algolia/requester-node-http';
 import { createUserAgent, Request, RequestOptions } from '@algolia/transporter';
@@ -346,6 +356,13 @@ export default function algoliasearch(
 
         return initPersonalization()(clientOptions);
       },
+      getRecommendations,
+      getFrequentlyBoughtTogether,
+      getLookingSimilar,
+      getRecommendedForYou,
+      getRelatedProducts,
+      getTrendingFacets,
+      getTrendingItems,
     },
   });
 }
@@ -697,6 +714,15 @@ export type SearchClient = BaseSearchClient & {
    * @deprecated Use `initPersonalization` instead.
    */
   readonly initRecommendation: (options?: InitPersonalizationOptions) => PersonalizationClient;
+  readonly getRecommendations: WithRecommendMethods<BaseSearchClient>['getRecommendations'];
+  readonly getFrequentlyBoughtTogether: WithRecommendMethods<
+    BaseSearchClient
+  >['getFrequentlyBoughtTogether'];
+  readonly getLookingSimilar: WithRecommendMethods<BaseSearchClient>['getLookingSimilar'];
+  readonly getRecommendedForYou: WithRecommendMethods<BaseSearchClient>['getRecommendedForYou'];
+  readonly getRelatedProducts: WithRecommendMethods<BaseSearchClient>['getRelatedProducts'];
+  readonly getTrendingFacets: WithRecommendMethods<BaseSearchClient>['getTrendingFacets'];
+  readonly getTrendingItems: WithRecommendMethods<BaseSearchClient>['getTrendingItems'];
 } & Destroyable;
 
 export * from '../types';